refactor: 린트룰에 맞춰 파일명을 대문자로 변경
